Add default impls for Run trait conversion fns
The Run trait has 3 methods to help with converting from a concrete type
to Box<dyn Any>,  &dyn Any and &mut dyn Any. These methods are
implemented for all types that implement Run, but the impls all just
return self. This commit adds default impls for these methods by
changing the trait to require that implementations which call these
methods are sized. This works for every type that currently implements
Run, and allows us to simplify the trait impls.